VisionFive2 launchpad added.
MACHINE

The VisionFive2 has many peripherals besides those implemented by QEMU
for the SiFive-U and it is a rather complex and powerful SoC, but, so
far, it matches the SiFive-U in all aspects regarding to the RV64 cores.

UART

VisionFive 2 supposedly uses a Synopsys DW8250 UART, which is register
compatible with the NS16500A. I therefore just expanded the offsets to
32-bit and created an initial hardware mediator for it. Config is
disabled because we don't know the clocks yet, but it works with the
default 115200 8N1.

PMP

Apparently, the machine does not require a PMP configuration to boot.